// src/store/enums/InventoryActionsEnum.ts
export enum InventoryActionTypes {
  GET_INVENTORIES = "GET_INVENTORIES",
  GET_INVENTORIES_SUCCESS = "GET_INVENTORIES_SUCCESS",
  GET_INVENTORIES_FAILURE = "GET_INVENTORIES_FAILURE",
  CREATE_INVENTORY = "CREATE_INVENTORY",
  CREATE_INVENTORY_SUCCESS = "CREATE_INVENTORY_SUCCESS",
  CREATE_INVENTORY_FAILURE = "CREATE_INVENTORY_FAILURE",
  UPDATE_INVENTORY = "UPDATE_INVENTORY",
  UPDATE_INVENTORY_SUCCESS = "UPDATE_INVENTORY_SUCCESS",
  UPDATE_INVENTORY_FAILURE = "UPDATE_INVENTORY_FAILURE",
  DELETE_INVENTORY = "DELETE_INVENTORY",
  DELETE_INVENTORY_SUCCESS = "DELETE_INVENTORY_SUCCESS",
  DELETE_INVENTORY_FAILURE = "DELETE_INVENTORY_FAILURE",
  ADD_INVENTORY = "ADD_INVENTORY",
  ADD_INVENTORY_SUCCESS = "ADD_INVENTORY_SUCCESS",
  ADD_INVENTORY_FAILURE = "ADD_INVENTORY_FAILURE",
  GET_INVENTORY_BY_NAME_OR_REFERENCE = "GET_INVENTORY_BY_NAME_OR_REFERENCE",
  GET_INVENTORY_BY_NAME_OR_REFERENCE_SUCCESS = "GET_INVENTORY_BY_NAME_OR_REFERENCE_SUCCESS",
  GET_INVENTORY_BY_NAME_OR_REFERENCE_FAILURE = "GET_INVENTORY_BY_NAME_OR_REFERENCE_FAILURE",
  GET_INVENTORY_BY_NAME_OR_REFERENCE_FILTER = "GET_INVENTORY_BY_NAME_OR_REFERENCE_FILTER",
  GET_INVENTORY_BY_NAME_OR_REFERENCE_SUCCESS_FILTER = "GET_INVENTORY_BY_NAME_OR_REFERENCE_SUCCESS_FILTER",
  GET_INVENTORY_BY_NAME_OR_REFERENCE_FAILURE_FILTER = "GET_INVENTORY_BY_NAME_OR_REFERENCE_FAILURE_FILTER",
  GET_INVENTORY_TO_REPORT = "GET_INVENTORY_TO_REPORT",
  GET_INVENTORY_TO_REPORT_SUCCESS = "GET_INVENTORY_TO_REPORT_SUCCESS",
  GET_INVENTORY_TO_REPORT_FAILURE = "GET_INVENTORY_TO_REPORT_FAILURE",
  GET_INVENTORIES_BY_CATEGORY_ID = "GET_INVENTORIES_BY_CATEGORY_ID",
  GET_INVENTORIES_BY_CATEGORY_ID_SUCCESS = "GET_INVENTORIES_BY_CATEGORY_ID_SUCCESS",
  GET_INVENTORIES_BY_CATEGORY_ID_FAILURE = "GET_INVENTORIES_BY_CATEGORY_ID_FAILURE",
}